feat: Long 타입으로 받기위해 수정 완료

+++ src/main/java/com/example/feedprep/domain/notification/service/NotificationService.java
@@ -17,5 +17,5 @@ public interface NotificationService {
 
 	 ApiResponseDto deleteNotification(Long userId, Long notificationId );
 
-	 Integer getNotificationCount(Long receiverId);
+	Long getNotificationCount(Long receiverId);
 }
